backend/requirements.txt
flask
flask_cors
currencyconverter
googletrans==4.0.0-rc1
PyPDF2
python-docx
google.generativeai
dotenv